Rename video room widget file to contain underscores
